Fix layout in hero and call to action section
One of the areas for improvement for me is how to work with both the
flexbox and traditional box model elements. A lot of the fixes to make
sure the layout aligns was by adding the right margin and padding.

Spend time looking into how `space-between`` and `space-around` can be
leveraged instead.

Also this looks so bad on mobile, looking forward to that section.